+++ packages/react/src/atomic/Button/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Button component allows us to initiate actions, make state or page changes.
 
+## 2025-11-13 `2.10.0`
+
+#### 🎉 New features
+
+- Added `large` size to `Button` component. ([#380](https://github.com/TiendaNube/nimbus-design-system/pull/380) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-10-17 `2.9.0`
 
 #### 🎉 New features